+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Bumps Instabug Android SDK to `v11.8.0`
 * Bumps Instabug iOS SDK to `v11.7.0`
+* Adds `BugReporting.setDisclaimerText` API
+* Adds `BugReporting.setCommentMinimumCharacterCount` API
 * Adds support for more locales:
   * czech
   * finnish